refactor: :recycle: Button refactor
