+++ libs/dana/src/components/Table/TableToolbarSearch.tsx
@@ -1,7 +1,7 @@
 import { SearchIcon } from '@dana-icons';
 import React, { InputHTMLAttributes, useState } from 'react';
 import { Props } from '../../helpers';
-import { Button } from '../forms/button';
+import { Button } from '../Button';
 import { Select, Option } from '../forms/select';
 import { TextInput } from '../TextInput';
 import { Row, Col } from '../layout';
