+++ libs/dana/src/components/index.ts
@@ -9,6 +9,7 @@ export * from './navigation';
 export * from './layout';
 
 export * from './Accordion';
+export * from './Button';
 export * from './Checkbox';
 export * from './fade-in';
 export * from './Image';
